Fix Qt6 compilation errors in qstring.h.
The complicated code in Qt6 to determine the type of an argument being
passed to QString::arg seems to have run afoul of changes to
libstdc++; changes similar to those the MythTV code fixed in
7c781b3860.  While trying to determine the type of a variable
(specifically an instance of a class), Qt6 checks to see if it can be
converted to an integer with an "operator int()" function and checks
to see if it can be converted to a string with a "toString()"
function.  On recently updated Fedora 43 and Rawhide systems, this
code is now throwing four fatal errors each time it tries to determine
if it can coerce a variable to a known type.  Its unclear if this is
because of an upgrade to Qt 6.10 or to libstdc++ 15.2.1.

Adding explicit calls to a "toInt()" function or explicitly casting
the variable to "int" isn't an elegant solution, but it solves the
compilation failures without having to wait for Qt to resolve the
problem.

(cherry picked from commit 87d57364e3c68344d178c42ad760bd145b21b7d3)

+++ mythtv/libs/libmythtv/cardutil.cpp
@@ -732,7 +732,7 @@ QString CardUtil::ProbeDVBType(const QString &device)
     ret = (type.toString() != "UNKNOWN") ? type.toString().toUpper() : ret;
 
     LOG(VB_GENERAL, LOG_DEBUG, LOC + QString("(%1) tuner type:%2 %3")
-        .arg(device).arg(type).arg(ret));
+        .arg(device).arg(type.toInt()).arg(ret));
 #endif // USING_DVB
 
     return ret;
@@ -837,7 +837,7 @@ DTVTunerType CardUtil::ConvertToTunerType(DTVModulationSystem delsys)
         default:
             LOG(VB_GENERAL, LOG_ERR, LOC +
                 QString("TODO Add to switch case delivery system:%2 %3")
-                    .arg(delsys).arg(delsys.toString()));
+                    .arg(delsys.toInt()).arg(delsys.toString()));
             break;
     }
 
@@ -934,7 +934,7 @@ DTVModulationSystem CardUtil::ProbeCurrentDeliverySystem([[maybe_unused]] const 
     delsys = ProbeCurrentDeliverySystem(fd_frontend);
 
     LOG(VB_GENERAL, LOG_DEBUG, QString("CardUtil(%1): delsys:%2 %3")
-        .arg(device).arg(delsys).arg(delsys.toString()));
+        .arg(device).arg(delsys.toInt()).arg(delsys.toString()));
 
     close(fd_frontend);
 #endif
